requirements.txt
discord.py
ollama
asyncio
